test(integration): consolidate sequence diagram suites into one Electron run
Sequence diagram and Compare sequence diagram both use the default project-a
workspace and identical setup. Merge them into one file so they share a single
Electron instance. Both already stub through per-suite sandboxes, so there is no
cross-suite state to leak.

Integration test files: 25 -> 24.
